Add intrinsics.hh, move cpuid tests from bits.hh
